feat: Add list models feature
This commit adds a new method `models()` to the `Groq` class, allowing users to retrieve a list of models. The method returns an instance of the `Models` class, which encapsulates the logic for making the API request and parsing the response.

The `models()` method is implemented in the `Groq` class and returns an instance of the `Models` class. The `Models` class has a `list()` method that sends a GET request to the `/models` endpoint and returns the response as an array.

This new feature enhances the functionality of the library by providing an easy way to retrieve and work with models.